requirements.txt
flask==2.3.3
ephem==4.1.4
Pillow==10.0.0
pytz==2023.3
numpy==1.24.3
hypothesis==6.82.6
pytest==7.4.0
pytest-cov==4.1.0
